requirements.txt
certifi==2018.11.29
chardet==3.0.4
clearbit==0.1.7
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
idna==2.7
pyhunter==0.3
PyJWT==1.6.4
pytz==2018.7
requests==2.20.1
selenium==3.141.0
urllib3==1.24.1
